veer: Add AXI FVIP
Use YosysHQ-GmbH/SVA-AXI4-FVIP to demo AXI properties in VeeR.  Specifically targets the DMA AXI with the `DESTINATION` protocol checker; with asserts on the outputs and assumes on the inputs.
Include (small) patch for `dma_ctrl.sv` to get the clock to work.
Add cover mode, as well as AXI targets to `sby` file.
Add `--keep-going` flag to bmc mode.